Solve "COMPUTE THE K CLOSEST STARS"
Use max-heap approach: start by pushing into the heap the first k stars; then
if you find a star that's smaller than the maximum, evict the maximum, insert
the new star. The idea is that if you have to keep a set of k elements and you
find a new one that you want to insert, the element that you should remove from
the set is the maximum already present in the set.